requirements.txt
fastapi>=0.115,<1
uvicorn>=0.32,<1
pandas>=2.2,<3
PyYAML>=6.0,<7
requests>=2.32,<3
openai>=1.65,<2
pytest>=8.3,<9
